requirements.txt
streamlit
sentence-transformers
faiss-cpu
PyMuPDF
pandas
numpy
requests
groq